chore: 优化代码,高dpi设备的支持移入pdfium库中
优化代码,高dpi设备的支持移入pdfium库中

Log: 优化代码,高dpi设备的支持移入pdfium库中

+++ 3rdparty/deepin-pdfium/src/dpdfannot.cpp
@@ -181,7 +181,7 @@ int DPdfLinkAnnot::pageIndex() const
 
 QPointF DPdfLinkAnnot::offset() const
 {
-    return QPointF(m_left, m_top);
+    return QPointF(static_cast<qreal>(m_left), static_cast<qreal>(m_top));
 }
 
 void DPdfLinkAnnot::setLinkType(int type)
